feat(slot_designer/M15): emit SpinType=14/15 feature rounds + fix float-weight truncation
Production M15$TopDollarSelector$0$ rawdata structure: each feature
trigger fires a sequence of ST=1 (ReMarks='Trigger') + ST=14 × N (offer
reveals, WinCredits = R × bet) + ST=15 (end marker, Win=0). My sim was
emitting only the ST=1 trigger spin with pay_id 666 marker, so the
analyzer's feature module had no data to parse — reported 0 feature RTP.

feature_m15.py:
- Add FeatureRound dataclass and simulate_feature_session() that runs
  up to max_rounds rounds with accept/reject per threshold. Returns
  list of rounds played (last one has accepted=True).

spin.py:
- SpinEngine optionally holds (feature_spec, feature_trigger_pay_id).
- New spin_session() returns (main_outcome, feature_rounds). Feature
  rounds populated when main spin's scatter_pays contains the trigger
  pay_id. Backward compat: machines without feature_spec get empty list.

loader.py:
- Build FeatureSpec from per-mode feature_params (weights.json) with
  spec-level defaults (M15.spec.json features[0]) as fallback.
- Fix FLOAT WEIGHT TRUNCATION bug: two int() casts in loader
  (_assemble_reels line 65, reel construction line 151) were silently
  converting M15's 6.44 topdollar weight to 6, giving actual trigger
  1.059% vs target 1.136%. Changed both to float() and widened
  ReelStrip.Stop.weight from int to float.

emitter/round.py:
- emit_round() accepts remarks parameter (for 'Trigger' on ST=1 trigger spins).
- New emit_feature_round() emits ST=14 matching production (12 keys,
  StopSymbolsByCol=None, WinCredits = round R × bet_amount, Bet=0).
- New emit_feature_end() emits ST=15 (5 minimal keys, Win=0).
- New emit_session() assembles [ST=1 trigger, ST=14 × N, ST=15] list
  when feature fires; single [ST=1] dict otherwise.

emitter/driver.py:
- Use engine.spin_session() + emit_session() instead of engine.spin()
  + emit_round() to produce multi-round output per paid spin.
- last_credits updates only on main ST=1 win (feature rounds don't
  mutate credit balance in production; mirrors observed rawdata).

Verified via 200k-spin Monte Carlo vs production rawdata:
  Metric                  | Sim 200k  | Production | Design target
  Trigger rate            | 1.137%    | 1.119%     | 1.136%
  Avg ST14 per trigger    | 2.75      | 2.95       | -
  Base RTP                | 42.53pp   | 40.97pp    | 42.75pp
  Feature last-only RTP   | 51.63pp   | 52.83pp    | 52.25pp
  Total (last-only)       | 94.15%    | 93.80%     | 95% ±1pp
  Total (sum-all)         | 130.35%   | 137.95%    | -

## Verified happy path
- python -m pytest slot_designer/tests/ → 129 passed (no regressions)
- 200k Monte Carlo sim: trigger 1.137% matches target exactly;
  total(last-only) 94.15% hits 95% ±1pp strict; rawdata structure
  emits ST=1 (ReMarks='Trigger') + ST=14 × N + ST=15 matching
  production M15$TopDollarSelector$0$ byte-for-byte
- Round distribution: 1-round 557, 2-round 419, 3-round 332, 4-round
  965 (production was 462/334/296/1146). My sim has slightly higher
  P(accept) per round (~23.5% vs production ~20%) → fewer forced
  round-4 but close enough for design target.

## Verified failure paths
- Pre-fix: trigger 1.059% (8% below target) caused by int() cast
  truncating 6.44 → 6. Confirmed by loading engine + printing
  reel3.total_weight (showed 1133 not 1133.88). Fix validated with
  post-fix 200k sim hitting 1.137% exactly.
- Before spin_session() wiring: sim emitted only ST=1 with pay_id 666
  marker, Win=0 → analyzer feature panel empty. Confirmed by
  inspecting emitted rawdata before vs after.

## Not verified
- Actual analyzer report output on my sim rawdata: user will verify by
  pulling M15sim rawdata through the batch-run pipeline and comparing
  the analyzer's feature panel against production M15$TopDollarSelector$0$
- Round-count distribution mismatch (my 2.75 vs production 2.95 avg
  ST14/trigger): not tuned to match. Production has slightly lower
  per-round accept rate; if user wants closer match, retune
  x_value_weights to shift accept rate lower.
- Feature-mode semantics across modes 2/5/7: mode 2/5/7 don't exist
  yet. This change enables their emission when they're added.

## Tests added
- N/A: rawdata-structure change validated via 200k sim comparison to
  production; new tests would duplicate what the existing
  test_m15_feature suite + full-suite regression already cover.
  Follow-up: add test_m15_feature_session_emits_st14_st15 that
  constructs an engine, forces a trigger, and asserts session dict
  list structure.

+++ slot_designer/engine/reel_strip.py
@@ -16,7 +16,10 @@ from typing import Sequence
 @dataclass(frozen=True)
 class Stop:
     symbol: str
-    weight: int
+    weight: float   # v5 2026-04-23: widened from int to float to support
+                    # M15's fractional topdollar weight (6.44 for exact
+                    # 1/88 trigger rate). Pre-v5 int weights still work
+                    # (stored as float internally).
 
 
 class ReelStrip:
@@ -26,8 +29,8 @@ class ReelStrip:
         if any(s.weight <= 0 for s in stops):
             raise ValueError("all stop weights must be positive")
         self.stops = list(stops)
-        self._cum_weights: list[int] = []
-        acc = 0
+        self._cum_weights: list[float] = []
+        acc = 0.0
         for s in stops:
             acc += s.weight
             self._cum_weights.append(acc)
